# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USC App Degree Projects API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Status"
            }
          },
          {
            "name": "modality_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality Id"
            }
          },
          {
            "name": "academic_period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Period"
            }
          },
          {
            "name": "academic_program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Program Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Inscribe una nueva idea de trabajo de grado (solo Estudiante).",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/sustentation": {
      "post": {
        "tags": [
          "sustentation"
        ],
        "summary": "Schedule Sustentation",
        "description": "El Administrador programa la sustentación pública.\nValida estado aprobado_para_sustentacion y que no sea modalidad Diplomado.\nTransiciona a sustentacion_programada y notifica a todos los actores.",
        "operationId": "schedule_sustentation_api_v1_projects__project_id__sustentation_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SustentationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SustentationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sustentation"
        ],
        "summary": "Get Sustentation",
        "description": "Retorna el detalle de la sustentación y sus calificaciones.\nPara Estudiante: juror_id oculto (respuesta anónima).",
        "operationId": "get_sustentation_api_v1_projects__project_id__sustentation_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SustentationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/sustentation/evaluations": {
      "post": {
        "tags": [
          "sustentation"
        ],
        "summary": "Submit Sustentation Evaluation",
        "description": "El jurado registra su calificación de la sustentación.\nSolo puede calificar el docente asignado como jurado en stage='sustentacion',\no el Administrador en su lugar.\nCuando ambos jurados han calificado: calcula final_score, determina is_approved\ny transiciona el estado del proyecto.",
        "operationId": "submit_sustentation_evaluation_api_v1_projects__project_id__sustentation_evaluations_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SustentationEvalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SustentationEvalAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedProjectsResponse",
        "example": {
          "items": [],
          "page": 1,
          "size": 20,
          "total": 0
        }
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "maxLength": 200,
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "member_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Member Ids"
          },
          "prerequisite_declaration": {
            "type": "boolean",
            "title": "Prerequisite Declaration"
          }
        },
        "type": "object",
        "required": [
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "member_ids",
          "prerequisite_declaration"
        ],
        "title": "ProjectCreate",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "member_ids": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          ],
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "prerequisite_declaration": true,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC"
        }
      },
...
      "SustentationCreate": {
        "properties": {
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "scheduled_time": {
            "type": "string",
            "title": "Scheduled Time"
          },
          "location": {
            "type": "string",
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "scheduled_date",
          "scheduled_time",
          "location"
        ],
        "title": "SustentationCreate",
        "example": {
          "location": "Sala de conferencias B, Edificio de Ingenierías",
          "scheduled_date": "2026-04-25",
          "scheduled_time": "09:30"
        }
      },
...
      "SustentationEvalCreate": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "SustentationEvalCreate",
        "example": {
          "score": 4.2
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}